Begin de-cluttering ToolsUI (3).
